app/requirements.txt
quart
aiogram
aiohttp
pillow
pyzbar
hypercorn
aiosqlite